# api.yml
openapi: 3.0.0
info:
  title: Edwin Perqara Interview Test (Storage Drink)
  version: 0.1.0
servers:
  - url: http://localhost

paths:
  /api/v1/storage/drinks:
    get:
      summary: Get all drinks inside of storage
      responses:
        '200':
          description: Success get list of drinks
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The id of drink
                  name:
                    type: string
                    description: The name of drink
                  stock:
                    type: integer
                    description: The stock of drink
        '400':
          description: Failed to retrieve list of drinks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/storage/drink:
    post:
      summary: Insert a new drink
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpec"
      responses:
        "201":
          description: Drink has been successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/storage/drink/{id}:
    put:
      summary: Update an existing of drink
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpec"
      responses:
        "200":
          description: "Drink has been successfully updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "Drink has been successfully deleted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
                example: "Drink is not exist"

components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateSpec:
      required:
        - "name"
        - "stock"
      type: object
      properties:
        name:
          description: The name of drink
          type: string
        stock:
          description: The stock of drink
          type: integer
          minimum: 1
    UpdateSpec:
      required:
        - "name"
        - "stock"
      type: object
      properties:
        name:
          description: The name of drink
          type: string
        stock:
          type: integer
          description: Adjust amount of stock
          minimum: 1
    SuccessResponse:
      properties:
        id:
          type: string
          description: The id of drink
          example: b5fe3aae-13fe-42df-a427-878678444dca
    ErrorResponse:
      properties:
        message:
          type: string
          description: error message
          example: drink is not exist
  